Novel 필드 오타 수정, 타입 변경, 선호작 정보 dto 생성

+++ src/main/java/com/brad/novel/preference/entity/Preference.java
@@ -24,6 +24,6 @@ public class Preference extends BaseEntity {
     @ManyToOne(fetch = LAZY)
     private Novel novel;
 
-    private int recentCh;  // 가장 마지막으로 읽은 화
-    private int likeNumber;      // 좋아요(선호) 개수
+    private Long recentCh;  // 가장 마지막으로 읽은 화
+    private Long likeNumber;      // 좋아요(선호) 개수
 }
